pyproject.toml
[project]
name = "wikimark-app"
version = "0.1.0"
description = "FastAPI application with GenAI and Database support"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.124.2",
    "uvicorn==0.38.0",
    "sqlalchemy==2.0.45",
    "psycopg2-binary==2.9.11",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.5.0",
    "langchain-google-genai==4.0.0",
    "wikipedia==1.4.0",
    "asyncpg>=0.31.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "email-validator>=2.3.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
